feat: atualização do componente

+++ lib/model/carro_model.dart
@@ -4,17 +4,16 @@ import 'package:carroeletrico/model/enum/controle_trava.dart';
 class Carro {
   int? chassi;
   int? porcentagemCarga;
-  double? quilometragem;
+  int? quilometragem;
   ControleStatus controleStatus;
   ControleTrava controleTrava;
   String apelido;
 
-  Carro({
-    this.chassi,
-    this.porcentagemCarga,
-    this.quilometragem,
-    this.controleStatus = ControleStatus.Desligado,
-    this.controleTrava = ControleTrava.Travado,
-    required this.apelido
-  });
+  Carro(
+      {this.chassi,
+      this.porcentagemCarga,
+      this.quilometragem,
+      this.controleStatus = ControleStatus.Desligado,
+      this.controleTrava = ControleTrava.Travado,
+      required this.apelido});
 }
